Rename ModeToggle and create more reusable Tailwind classes
